Update README.md
The story follows the lives of Ae-sun and Gwan-sik in the 1950s as they navigate love and family. The slice of drama is a heartwarming depiction of first love, reminiscent of old-school love. The show is set against the beautiful backdrop of Jeju Island and will leave your heart full of emotions